Added Result Object as response and changed datatype to Long for ids

+++ src/main/java/com/turvo/abcbanking/dao/BranchManagementDAO.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author <a href="sampath.rachakonda@imaginea.com">srachakonda</a>
  * @version $Revision: 1.0$, $Date: 20-Feb-2018
  */
-public interface BranchManagementDAO extends JpaRepository<Branch, Integer> {
+public interface BranchManagementDAO extends JpaRepository<Branch, Long> {
 }
