+++ src/main/java/com/turvo/abcbanking/dao/CounterManagementDAO.java
@@ -10,7 +10,7 @@ import java.util.List;
  * @author <a href="sampath.rachakonda@imaginea.com">srachakonda</a>
  * @version $Revision: 1.0$, $Date: 12-Feb-2018
  */
-public interface CounterManagementDAO extends JpaRepository<Counter, Integer> {
+public interface CounterManagementDAO extends JpaRepository<Counter, Long> {
 
     List<Counter> findByAccountType(AccountType accountType);
